Add GropChangeColor.java, GroupMaximumAugmentation.java

+++ src/modules/GroupDepreciation.java
@@ -33,8 +33,8 @@ public class GroupDepreciation extends SentenceGenerator{
         Session[] sessions = filter.filter(Input.inputData[stockExchange].getSessions(), group);
         String ticker1 = new String();
         String ticker2 = new String();
-        float percentReduction1 = 0, percentReduction2 = 0;
-        Map<String, Float> map = new HashMap<>();
+        double percentReduction1 = 0, percentReduction2 = 0;
+        Map <String, Float> map = new HashMap<>();
 
         Date today = sessions[0].getDate();
         Date previousDay = null; // Ngày giao dịch trước đó
